requirements.txt
appdirs==1.4.4
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
backcall==0.1.0
black==19.10b0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
city-scrapers-core==0.5.3
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
hyperlink==18.0.0
icalendar==4.0.3
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
legistar==0.0.1
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.782
mypy-extensions==0.4.3
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pathspec==0.8.0
pbr==5.4.3
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.4.3
queuelib==1.5.0
regex==2020.5.14
requests==2.22.0
scipy==1.2.1
scrapelib==1.2.0
Scrapy==1.6.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
stevedore==1.31.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.1
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.2
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uber-rides==0.6.0
urllib3==1.25.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
zope.interface==4.6.0
ics